backend/requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
charset-normalizer==3.4.2
django==4.2.23
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==10.4.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.9.0
reportlab==4.4.2
six==1.17.0
sqlparse==0.5.3
typing-extensions==4.13.2
